+++ extension-al-object-designer/src/ALSettings.ts
@@ -9,8 +9,17 @@ export class ALSettings {
         this._settings = (config.get('configurations') as Array<any>)[0];
     }
 
-    public get(key: string): any
-    {
+    public get(key: string): any {
         return this._settings[key] || '';
+    }
+
+    public getUrl(type: string, id: number): string {
+        let server = this.get('server'),
+            instance = this.get('serverInstance');
+
+        server = server[server.length - 1] != '/' ? server + '/' : server;
+        let launchUrl = `${server}${instance}/?${type.toLowerCase()}=${id}`;
+
+        return launchUrl;
     }
 }